Get_tallent/requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
